add Func Random

+++ foundry/src/sol-004.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.20;
 
 contract StateVariables {
-    uint public num;
+    uint256 public num;
 
-    function setNum(uint _num) external {
+    function setNum(uint256 _num) external {
         num = _num;
     }
 
@@ -15,11 +15,11 @@ contract StateVariables {
     // What is "view"?
     // "view" tells Solidity that this is a read-only function.
     // It does not make any updates to the blockchain.
-    function getNum() external view returns (uint) {
+    function getNum() external view returns (uint256) {
         return num;
     }
 
-    function getNumPlusOne() external view returns (uint) {
+    function getNumPlusOne() external view returns (uint256) {
         return num + 1;
     }
 }
