+++ foundry/src/sol-010.sol
@@ -2,7 +2,6 @@
 pragma solidity ^0.8.20;
 
 contract Constants {
-    address public constant MY_ADDR =
-        0x777788889999AaAAbBbbCcccddDdeeeEfFFfCcCc;
-    uint public constant MY_UINT = 123;
+    address public constant MY_ADDR = 0x777788889999AaAAbBbbCcccddDdeeeEfFFfCcCc;
+    uint256 public constant MY_UINT = 123;
 }
